serverless/entity-extract-serverless/extract/requirements.txt
requests
pydantic>=2.5
amazon-textract-textractor
datamodel-code-generator
pyyaml